+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/ConsumerCoordinator.java
@@ -71,7 +71,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
     private final SubscriptionState subscriptions;
     private final OffsetCommitCallback defaultOffsetCommitCallback;
     private final boolean autoCommitEnabled;
-    private final int autoCommitIntervalMs;
+    private final long autoCommitIntervalMs;
     private final ConsumerInterceptors<?, ?> interceptors;
     private final boolean excludeInternalTopics;
     private final AtomicInteger pendingAsyncCommits;
@@ -103,7 +103,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
                                Time time,
                                long retryBackoffMs,
                                boolean autoCommitEnabled,
-                               int autoCommitIntervalMs,
+                               long autoCommitIntervalMs,
                                ConsumerInterceptors<?, ?> interceptors,
                                boolean excludeInternalTopics,
                                final boolean leaveGroupOnClose) {
